Fix: Fixed the bug with instruction cache.
Details: The situation considered is when branch instruction is followed by an invalid instruction, it starts caching new instructions. If the branch is not taken, everything works fine, but the
problem araises if the branch is taken. In that situation branch flushes execute and decode unit pipeline registers and resets branch flag. Therefore after instruction caching is finished the branch
that should have been takes will be ignored. In order to solve the issue, for branch flushes additional condition was added to check if there is no caching going on.